pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "joblib>=1.4.2",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "plotly>=5.24.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv",
    "scikit-learn>=1.5.2",
    "sqlalchemy>=2.0.36",
    "streamlit>=1.40.1",
]
